+++ src/routes/v2/projects/[projectId]/workflows/[workflowId]/+page.svelte
@@ -10,7 +10,6 @@
 	import Modal from '$lib/components/common/Modal.svelte';
 	import StandardDismissableAlert from '$lib/components/common/StandardDismissableAlert.svelte';
 	import VersionUpdate from '$lib/components/v2/workflow/VersionUpdate.svelte';
-	import { getAllNewVersions } from '$lib/components/v2/workflow/version-checker';
 	import ImagesStatus from '$lib/components/jobs/ImagesStatus.svelte';
 	import TasksOrderModal from '$lib/components/v2/workflow/TasksOrderModal.svelte';
 	import { extractRelevantJobError } from '$lib/common/job_utilities';
@@ -105,10 +104,8 @@
 	/** @type {TypeFiltersFlowModal|undefined} */
 	let typeFiltersFlowModal = $state();
 
-	/** @type {{ [id: string]: import('fractal-components/types/api').TaskV2[] }} */
+	/** @type {{ [id: string]: Array<{ task_id: number, version: string }> }} */
 	let newVersionsMap = $state({});
-	/** @type {{ [id: string]: string | null }} */
-	let tasksVersions = $state({});
 
 	/** @type {import('fractal-components/types/api').ApplyWorkflowV2|undefined} */
 	let selectedSubmittedJob = $state();
@@ -460,13 +457,33 @@
 	}
 
 	async function checkNewVersions() {
-		if (workflow) {
-			const { updateCandidates, enrichedTasks } = await getAllNewVersions(
-				workflow.task_list.map((wt) => wt.task)
-			);
-			newVersionsMap = updateCandidates;
-			tasksVersions = Object.fromEntries(enrichedTasks.map((t) => [t.id, t.version]));
+		if (!workflow) {
+			return;
 		}
+		newVersionsMap = await getNewVersionsForWorkflow(workflow);
+	}
+
+	/**
+	 * @param {import('fractal-components/types/api').WorkflowV2} workflow
+	 * @returns {Promise<{ [id: string]: Array<{ task_id: number, version: string }> }>}
+	 */
+	async function getNewVersionsForWorkflow(workflow) {
+		const response = await fetch(
+			`/api/v2/project/${workflow.project_id}/workflow/${workflow.id}/version-update-candidates`
+		);
+
+		/** @type {Array<Array<{ task_id: number, version: string }>>} */
+		const updates = await response.json();
+
+		/** @type {{ [id: string]: Array<{ task_id: number, version: string }> }} */
+		const updateCandidates = {};
+
+		for (let i = 0; i < updates.length; i++) {
+			const task = workflow.task_list[i].task;
+			updateCandidates[task.id] = updates[i];
+		}
+
+		return updateCandidates;
 	}
 
 	let newVersionsCount = $state(0);
@@ -1059,10 +1076,7 @@
 							<div id="info-tab" class="tab-pane show active">
 								<div class="card-body">
 									{#if selectedWorkflowTask}
-										<TaskInfoTab
-											task={selectedWorkflowTask.task}
-											taskVersion={tasksVersions[selectedWorkflowTask.task.id]}
-										/>
+										<TaskInfoTab task={selectedWorkflowTask.task} />
 									{/if}
 								</div>
 							</div>
@@ -1088,6 +1102,7 @@
 									<VersionUpdate
 										workflowTask={selectedWorkflowTask}
 										updateWorkflowCallback={taskUpdated}
+										updateCandidates={newVersionsMap[selectedWorkflowTask.task_id] || []}
 										{updateNewVersionsCount}
 									/>
 								{/if}