+++ CHANGELOG.md
@@ -1,5 +1,9 @@
 *Note: Numbers like (\#123) point to closed Pull Requests on the fractal-web repository.*
 
+# Unreleased
+
+* Used new API endpoint to detect task version upgrade (\#792);
+
 # 1.18.0
 
 * Upgrade to Svelte 5 (\#789);